% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R
\name{accToComp}
\alias{accToComp}
\title{Convert Raw Accelerometer Data to Compliance Data}
\usage{
accToComp(fileName1, fileName2 = NULL, cut24 = FALSE, attachHours = FALSE)
}
\arguments{
\item{fileName1}{Path to first input file or filename if in current directory}

\item{fileName2}{Path/name of second input file if present. DEFAULT NULL}

\item{cut24}{boolean varable. True will cut the data at 24 hours before returning. DEFAULT FALSE}

\item{attachHours}{boolean varable. True will include number of raw hours recorded in returned value. DEFALUT FALSE}
}
\value{
A vector of 1s and 0s to indicate compliance and non-compliance. if attachHours = TRUE; list with 2 elements: 1. raw hours recorede, 2. comp data
}
\description{
This function will convert Raw Accelerometer Data to Compliance Data
using all of the accelerometer functions using their default arguements
}
