% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R, R/load_acc.R
\name{loadAccData}
\alias{loadAccData}
\alias{loadAccData}
\title{Load Accelerometer Data}
\usage{
loadAccData(fileName1, fileName2 = NULL, printHrs = FALSE)

loadAccData(fileName1, fileName2 = NULL, printHrs = FALSE)
}
\arguments{
\item{fileName1}{Path to first input file or filename if in current directory}

\item{fileName2}{Path to second input file or filename if in current directory. DEFAULT NULL}

\item{printHrs}{boolean varable. True will print time recorded to concle. DEFAULT FALSE}

\item{fileName1}{Path to first input file or filename if in current directory}

\item{fileName2}{Path to second input file or filename if in current directory. DEFAULT NULL}

\item{printHrs}{boolean varable. True will print time recorded to concle. DEFAULT FALSE}
}
\value{
A dataframe with 4 columns: 1. time stamps and 2., 3., 4. raw accelerometer data on 3 axis

A dataframe of the raw accelerometer data on 3 axis and time stamps
}
\description{
This function loads in accelerometer data.  It cleans the raw CSV files and
combines multiple files if needed. It assumes that the data is in
1 or 2 .CSV files in the specific format that comes from the accelerometer.
Also assumes that fileName1 is data recoreded before fileName2. If there is
only one file, fileName2 and printHrs can be omited.

This function loads in accelerometer data.  It cleans the raw CSV files and
combines multiple files if needed. It assumes that the data is in
1 or 2 .CSV files in the specific format that comes from the accelerometer.
Also assumes that fileName1 is data recoreded before fileName2. If there is
only one file, fileName2 and printHrs can be omited.
}
