% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R
\name{makeCompliance}
\alias{makeCompliance}
\title{Convert Movement Data to Compliance data}
\usage{
makeCompliance(movement, threshold = 400)
}
\arguments{
\item{movement}{A vector of smooth movement points such as that returned by makeMovement}

\item{threshold}{Septeration point btw compliant & noncompliant. DEFAULT 400}
}
\value{
A vector of 1s and 0s to indicate compliance
}
\description{
This function converts movement data to compliance binary data.
1 means the subject was wearing the bag at that minute, 0 means not.
Thus, each point is defined as compliant - 1, or noncompliant - 0.
The threshold variable set the level of septeration between which
points are compliant and which are noncompliant.
}
\details{
NOTE: Various threshold testing was performed to select the optimal level of 400
}
