% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R
\name{makeMovement}
\alias{makeMovement}
\title{Convert Vector Magnitude to Movement Data}
\usage{
makeMovement(vm3, rollLength = 5)
}
\arguments{
\item{vm3}{A vector of data points such as that returned by vecMag}

\item{rollLength}{Length of data to apply on rolling std dev. DEFAULT 5}
}
\value{
'movement' - A vector of smooth movement data points
}
\description{
This function converts vector magnitude data to movement data.
Vector magnitude data is the data returned by the vecMag function.
The functions applys a rolling standard deviation to the data input
to smooth the data and more closely resemble continuous movement.
}
