% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R
\name{switchCompliance}
\alias{switchCompliance}
\title{Change Sequences of Compliance Data to Opposite Classification}
\usage{
switchCompliance(mv, seqLen = 10, nonToComp = TRUE)
}
\arguments{
\item{mv}{List of accel data indicating compliant or non-compliant}

\item{seqLen}{Minimum number of acceptable sequeciencal points. DEFAULT 10}

\item{nonToComp}{Boolean indicating change of non-comp to comp - TRUE or comp to non-comp - FALSE}
}
\value{
A vector of 1s and 0s without sequences of 1s or 0s smaller than 'seqLen' argument
}
\description{
This function will convert non-compliant points to compliant points
or vice versa.  It is used by the switchFake function to further increase
the accuracy of the data on each subject.  Thus, the data is further smoothed.
}
\details{
Motivation: a subject might set the bag down for a few minutes and cause
the data to be non-compliant.  If the bag is only set down for a small period
of time, we dont want that period to be misclassified as non-compliant.
A similar situation exists for misclassifying compliant periods.
}
