% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R
\name{switchFake}
\alias{switchFake}
\title{Remove Misclassified Compliance Data Points}
\usage{
switchFake(mv, seqLen = 10, nonCompFirst = TRUE)
}
\arguments{
\item{mv}{List of accel data indicating compliant or non-compliant}

\item{seqLen}{Minimum number of acceptable sequeciencal points. DEFAULT 10}

\item{nonCompFirst}{Boolean indicating order of switching compliance type}
}
\value{
A vector of 1s and 0s without sequences of 1s AND 0s smaller than 'seqLen' argument
}
\description{
This function will convert misclassified non-compliant points to compliant
points and misclassified compliant points to non-compliant.
It is uses the switchCompliance function twice: once with nonToComp = TRUE &
once with nonToComp = FALSE.  nonToComp = TRUE is preformed first to error
on the side of compliance but this can be controled by the 'nonCompFirst' boolean
}
\details{
Motivation: a subject might set the bag down for a few minutes and cause
the data to be non-compliant.  If the bag is only set down for a small period
of time, we dont want that period to be misclassified as non-compliant.
A similar situation exists for misclassifying compliant periods.
}
