% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_fxns.R
\name{vecMag}
\alias{vecMag}
\title{Convert Accelerometer Data to Vector Magnitude}
\usage{
vecMag(acceldata)
}
\arguments{
\item{acceldata}{Dataframe object returned by loadAccData function}
}
\value{
'vm3' - A vector 1 minute data points.
}
\description{
This function converts acceldata to 1-minute vecter magnatude data.
This makes the data much lighter and easier to work with or plot.
The function takes 3 steps: 1. Vector magnitude, calculatd as square root
of sum of squares, 2. Vector magnitude, after subtracting out mode and
taking abs value, & 3. convert to 1-minute block average.
Each data point of the data returned represents 1 minute of raw acc data.
}
